requirements.txt.txt
It is required to install pyjwt, flask, sqlalchemy, psycopg2 before using this code
webencodings==0.5.1
flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.25
flask==2.0.1
pyjwt==2.2.0
psycopg2==2.9.1
requests==2.26.0